refactor: Combine integration test crates
